Following is working.....
Almost. It's not taking the correct event group yet, just the first from the json, and registration isn't behaving nicely.

Need to dtyle follow button and also fix some shit but hey I set out to do it and got something out there.

Too late for the college project but this app can be more than that.